+++ app/src/main/java/com/google/firebase/example/fireeats/ui/main/MainViewModel.java
@@ -6,7 +6,7 @@ import android.arch.lifecycle.Transformations;
 import android.arch.lifecycle.ViewModel;
 
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.example.fireeats.Resource;
+import com.google.firebase.example.fireeats.common.Resource;
 import com.google.firebase.example.fireeats.model.Restaurant;
 import com.google.firebase.example.fireeats.repo.Repository;
